Refactor: modified package of webclient
