+++ src/main/java/com/zerobase/hseungho/restaurantreservation/global/adapter/webclient/dto/CoordinateDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.hseungho.restaurantreservation.global.webclient.dto;
+package com.zerobase.hseungho.restaurantreservation.global.adapter.webclient.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
